Risk of Renal Stone Data for Cochran-Armitage Trend Test

Presents the frequencies of individuals at high risk of renal calculi according to age and gender. This a demo dataset for Cochran-Armitage trend test for investigating whether there is a linear trend between the proportion of individual with renal stone and ages.

It is the data set used in the Datanovia tutorial “Cochran-Armitage Trend Test in R: Test a Trend in Proportions” (https://www.datanovia.com/learn/biostatistics/categorical/cochran-armitage-trend-test-in-r).

data("renalstone")

Format

A data frame with 3513 rows and 3 columns (stored as a tibble).

gender

the individual's gender, "male" or "female".

stone

whether the individual has a renal stone, "yes" or "no".

age

the age group, an ordered factor: "30-39" < "40-49" < "50-59".

Source

Frequencies reported by Hazra and Gogtay (2016), expanded to one row per individual.

References

Hazra, Avijit, and Nithya Jaideep Gogtay. 2016. "Biostatistics Series Module 4: Comparing Groups - Categorical Variables." In Indian Journal of Dermatology.

See also

Examples

data(renalstone)
xtabs(~stone+age+gender, data = renalstone)
#> , , gender = male
#> 
#>      age
#> stone 30-39 40-49 50-59
#>   yes   188   313   232
#>   no    348   383   206
#> 
#> , , gender = female
#> 
#>      age
#> stone 30-39 40-49 50-59
#>   yes   196   223   103
#>   no    603   486   232
#>