Stress Data for Two-Way ANCOVA

Researchers want to evaluate the effect of a new "treatment" and "exercise" on the stress score reduction after adjusting for "age".

Two-way ANCOVA can be performed in order to determine whether there is interaction between exercise and treatment on the stress score.

data("stress")

Format

A data frame with 60 rows and 5 columns.

Examples

data(stress)
head(as.data.frame(stress))

#>   id score treatment exercise age
#> 1  1  95.6       yes      low  59
#> 2  2  82.2       yes      low  65
#> 3  3  97.2       yes      low  70
#> 4  4  96.4       yes      low  66
#> 5  5  81.4       yes      low  61
#> 6  6  83.6       yes      low  65