Heart Attack Data for Three Way ANOVA

Measures of cholesterol concentration in 72 participants treated with three different drugs. The aim is to examine the potential of new class of drugs in lowering the cholesterol concentration and consequently reducing heart attack.

The participants include 36 males and 36 females. Males and females were further (equally) subdivided into whether they were at low or high risk of heart attack.

This data set is suited for three way Anova test.

It contain the following variables:

gender, which has two categories: "male" and "female";

risk which has two levels: "low" and "high"

drug, which has three categories: "A", "B" and "C".

data("heartattack")

Format

A data frame with 72 rows and 5 columns.

Examples

data(heartattack)
head(as.data.frame(heartattack))

#>   gender risk drug cholesterol id
#> 1   male  low    A    5.238730  1
#> 2   male  low    A    5.075693  2
#> 3   male  low    A    4.678653  3
#> 4   male  low    A    5.361076  4
#> 5   male  low    A    4.957381  5
#> 6   male  low    A    4.828897  6