Weight Loss Score Data for Three-way Repeated Measures ANOVA

A researcher wanted to assess the effects of Diet and Exercises on weight loss in 10 sedentary males.

The participants were enrolled in four trials: (1) no diet and no exercises; (2) diet only; (3) exercises only; and (4) diet and exercises combined.

Each participant performed all four trials. The order of the trials was counterbalanced and sufficient time was allowed between trials to allow any effects of previous trials to have dissipated (i.e., a "wash out" period).

Each trial lasted nine weeks and the weight loss score was measured at the beginning of each trial (t1), at the midpoint of each trial (t2) and at the end of each trial (t3).

Three-way repeated measures ANOVA can be performed in order to determine whether there is interaction between diet, exercises and time on the weight loss score.

data("weightloss")

Format

A data frame with 48 rows and 6 columns.

Examples

data(weightloss)
head(as.data.frame(weightloss))

#>   id diet exercises    t1    t2    t3
#> 1  1   no        no 10.43 13.21 11.59
#> 2  2   no        no 11.59 10.66 13.21
#> 3  3   no        no 11.35 11.12 11.35
#> 4  4   no        no 11.12  9.50 11.12
#> 5  5   no        no  9.50  9.73 12.28
#> 6  6   no        no  9.50 12.74 10.43